+++ lang/src/test/java/net/openhft/lang/io/serialization/ExternalizableMarshallerTest.java
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package net.openhft.lang.io;
+package net.openhft.lang.io.serialization;
 
+import net.openhft.lang.io.NativeBytes;
 import org.junit.Test;
 import sun.nio.ch.DirectBuffer;
